movie_api/requirements.txt
asgiref==3.5.2
Django==4.0
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
